+++ flutter_web_auth_2/lib/src/flutter_web_auth_2_macos.dart
@@ -1,0 +1,1 @@
+// MacOS just needs the MethodChannel implementation