+++ flutter_web_auth_2_platform_interface/lib/flutter_web_auth_2_platform_interface.dart
@@ -1,4 +1,4 @@
-import 'package:flutter_web_auth_2_platform_interface/flutter_web_auth_2_method_channel.dart';
+import 'package:flutter_web_auth_2_platform_interface/method_channel/method_channel_flutter_web_auth_2.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
 /// The interface that implementations of FlutterWebAuth must implement.
@@ -7,19 +7,18 @@ import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 /// because `implements` does not consider newly added methods to be breaking
 /// changes. Extending this class (using `extends`) ensures that the subclass
 /// will get the default implementation.
-abstract class FlutterWebAuth2PlatformInterface extends PlatformInterface {
-  FlutterWebAuth2PlatformInterface() : super(token: _token);
-
-  static FlutterWebAuth2PlatformInterface _instance =
-      FlutterWebAuth2MethodChannel();
+abstract class FlutterWebAuth2Platform extends PlatformInterface {
+  FlutterWebAuth2Platform() : super(token: _token);
 
   static final Object _token = Object();
 
-  static FlutterWebAuth2PlatformInterface get instance => _instance;
+  static FlutterWebAuth2Platform _instance = FlutterWebAuth2MethodChannel();
+
+  static FlutterWebAuth2Platform get instance => _instance;
 
   /// Platform-specific plugins should set this with their own platform-specific
   /// class that extends [PlatformInterface] when they register themselves.
-  static set instance(FlutterWebAuth2PlatformInterface instance) {
+  static set instance(FlutterWebAuth2Platform instance) {
     PlatformInterface.verifyToken(instance, _token);
     _instance = instance;
   }
@@ -29,9 +28,11 @@ abstract class FlutterWebAuth2PlatformInterface extends PlatformInterface {
     required String callbackUrlScheme,
     required bool preferEphemeral,
   }) =>
-      throw UnimplementedError('authenticate() has not been implemented.');
-
-  Future clearAllDanglingCalls() => throw UnimplementedError(
-        'clearAllDanglingCalls() has not been implemented.',
+      _instance.authenticate(
+        url: url,
+        callbackUrlScheme: callbackUrlScheme,
+        preferEphemeral: preferEphemeral,
       );
+
+  Future clearAllDanglingCalls() => _instance.clearAllDanglingCalls();
 }