+++ flutter_web_auth_2_windows/example/main.dart
@@ -1,4 +1,0 @@
-void main() {
-  print('See flutter_web_auth_2 for more information:');
-  print('https://pub.dev/packages/flutter_web_auth_2/example');
-}